requirements.txt
aniso8601==8.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
